process self exit

+++ server.py
@@ -16,17 +16,20 @@ def run_web():
 web_process = Process(target=run_web)
 web_process.start()
 
-proc_pool = []
+proc_pool = {}
 
 def clean_up():
     print "stopping web process"
     web_process.terminate()
     web_process.join(5)
     print "stopping work process"
-    for p in proc_pool:
-        print "stopping process " + str(p.pid)
-        p.terminate()
-        p.join(5)
+    for (pid, p) in proc_pool:
+        if (p.is_alive()):
+            print "stopping process " + str(pid)
+            p.terminate()
+            p.join()
+        else:
+            p.join()
 
 atexit.register(clean_up)
 
@@ -35,18 +38,16 @@ def main():
         item = data_queue.get()
         print "data get " + item['type']
         if (item['type'] == 'webhook' and handler.can_handle(item['data'])):
-            proc = Process(target=handler.handle, args=(item['data'],))
+            proc = Process(target=handler.handle, args=(item['data'],data_queue))
             proc.start()
-            proc_pool.append(proc)
+            proc_pool[proc.pid] = proc
             print 'process {} started'.format(proc.pid)
-        time.sleep(0.1)
-        alive_pool = []
-        for p in proc_pool:
-            if (p.is_alive()):
-                alive_pool.append(p)
-            else:
-                p.join(5)
-        time.sleep(0.1)
+        elif (item['type'] == 'exit'):
+            end_proc = proc_pool.pop(item['pid'])
+            if (end_proc):
+                print 'process {} ended'.format(end_proc.pid)
+                end_proc.join()
+        time.sleep(0.3)
 
 if __name__ == "__main__":
     main()
